Added a module in the concerns folder to reduce duplicate code in conferences controllers
