+++ app/controllers/orga/conferences_controller.rb
@@ -1,13 +1,11 @@
 class Orga::ConferencesController < Orga::BaseController
   before_action :find_conference, only: [:show, :destroy]
 
+  include Conferencable
+
   def import
     Conference::Importer.call(params[:file].path, content_type: params[:file].content_type)
     redirect_to orga_conferences_path, notice: "Import finished! Check log for errors."
-  end
-
-  def index
-    @conferences = conferences
   end
 
   def destroy
@@ -21,10 +19,6 @@ class Orga::ConferencesController < Orga::BaseController
     @conference ||= Conference.find(params[:id])
   end
 
-  def conferences
-    Conference.ordered(sort_params).in_current_season
-  end
-
   def conference_params
     params.require(:conference).permit(
       :name, :location, :city, :country, :region,
@@ -35,13 +29,6 @@ class Orga::ConferencesController < Orga::BaseController
       :notes,
       conference_preferences_attributes: [:id, :_destroy]
     )
-  end
-
-  def sort_params
-    {
-      order: %w(name gid starts_on city country region).include?(params[:sort]) ? params[:sort] : nil,
-      direction: %w(asc desc).include?(params[:direction]) ? params[:direction] : nil
-    }
   end
 
   def set_breadcrumbs
